support: don't dump file in bind mounted folders (bsc#1243297)
Dumping the content of all the files in a bind-mounted folder can take a
lot of time and even be eating a lot of memory. Only dumping the
bind-mounted files in the supportconfig since those could contain
configuration tuning.
