+++ shared/utils/utils.go
@@ -262,82 +262,6 @@ func GetLocalTimezone() string {
 	return strings.TrimSpace(string(out))
 }
 
-// IsEmptyDirectory return true if a given directory is empty.
-func IsEmptyDirectory(path string) bool {
-	files, err := os.ReadDir(path)
-	if err != nil {
-		log.Fatal().Err(err).Msgf(L("cannot check content of %s"), path)
-		return false
-	}
-	if len(files) > 0 {
-		return false
-	}
-	return true
-}
-
-// RemoveDirectory remove a given directory.
-func RemoveDirectory(path string) error {
-	if err := os.Remove(path); err != nil {
-		return Errorf(err, L("Cannot remove %s folder"), path)
-	}
-	return nil
-}
-
-// FileExists check if path exists.
-func FileExists(path string) bool {
-	_, err := os.Stat(path)
-	if err == nil {
-		return true
-	} else if !os.IsNotExist(err) {
-		log.Fatal().Err(err).Msgf(L("Failed to get %s file informations"), path)
-	}
-	return false
-}
-
-// ReadFile returns the content of a file and exit if there was an error.
-func ReadFile(file string) []byte {
-	out, err := os.ReadFile(file)
-	if err != nil {
-		log.Fatal().Err(err).Msgf(L("Failed to read file %s"), file)
-	}
-	return out
-}
-
-// GetFileBoolean gets the value of a file containing a boolean.
-//
-// This is handy for files from the kernel API.
-func GetFileBoolean(file string) bool {
-	return strings.TrimSpace(string(ReadFile(file))) != "0"
-}
-
-// UninstallFile uninstalls a file.
-func UninstallFile(path string, dryRun bool) {
-	if FileExists(path) {
-		if dryRun {
-			log.Info().Msgf(L("Would remove file %s"), path)
-		} else {
-			log.Info().Msgf(L("Removing file %s"), path)
-			if err := os.Remove(path); err != nil {
-				log.Info().Err(err).Msgf(L("Failed to remove file %s"), path)
-			}
-		}
-	}
-}
-
-// TempDir creates a temporary directory.
-func TempDir() (string, func(), error) {
-	tempDir, err := os.MkdirTemp("", "mgradm-*")
-	if err != nil {
-		return "", nil, Error(err, L("failed to create temporary directory"))
-	}
-	cleaner := func() {
-		if err := os.RemoveAll(tempDir); err != nil {
-			log.Error().Err(err).Msg(L("failed to remove temporary directory"))
-		}
-	}
-	return tempDir, cleaner, nil
-}
-
 // GetRandomBase64 generates random base64-encoded data.
 func GetRandomBase64(size int) string {
 	data := make([]byte, size)
