+++ uyuni-tools.changes.cbosdo.support-files
@@ -1,0 +1,1 @@
+- support: don't dump files in bound folders (bsc#1243297)
